unify shaders

+++ ggml/src/ggml-vulkan/ggml-vulkan-types.h
@@ -1432,5 +1432,5 @@ struct ggml_backend_vk_device_context {
 };
 
 inline size_t ggml_vk_tensor_storage_unit_size(ggml_type type) {
-    return ggml_get_type_traits(type)->row_meta_size != 0 ? sizeof(uint16_t) : ggml_type_size(type);
+    return ggml_get_type_traits(type)->row_meta_size != 0 ? sizeof(uint8_t) : ggml_type_size(type);
 }
